Extending Table class
This is the only Actor of Stage in our case. Within this Actor we are creating 2 more actors i.e. Target and Stalker. That is the reason, we have to call super. act() and super. draw() method within its own act and draw method to call its child Actor's act and draw method. Very Important concept.